feat: move get range temp method
