+++ src/Views/Home/Home.tsx
@@ -32,6 +32,8 @@ import {
     formattedUmidity
 } from '../../helpers/format'
 
+import { getRangeTemp } from '../../helpers/range-temp'
+
 
 function Home() {
 
@@ -64,33 +66,7 @@ function Home() {
 
     }
 
-    const getRangeTemp = (list: any, sum: number) => {
 
-        const today = new Date()
-        const tomorrow = new Date(today.setDate(today.getDate() + sum))
-
-        const filteredList = list.filter((item: any) => {
-            return new Date(item.dt_txt).toDateString() == tomorrow.toDateString()
-        })
-
-
-        const maxTemp = filteredList.reduce(function (prev: any, current: any) {
-
-            return (prev.main.temp_max > current.main.temp_max) ? prev : current
-        })
-
-        const minTemp = filteredList.reduce(function (prev: any, current: any) {
-            return (prev.main.temp_min < current.main.temp_min) ? prev : current
-        })
-
-        const icon = filteredList[0].weather[0].icon
-
-        return {
-            min: minTemp.main.temp_min,
-            max: maxTemp.main.temp_max,
-            icon
-        }
-    }
 
     const getWeatherInformations = async (lat: number, long: number) => {
         try {
@@ -146,8 +122,6 @@ function Home() {
         }
     }
 
-
-
     const handleSubmit = async (event: React.SyntheticEvent): Promise<void> => {
         event.preventDefault()
         const { lat, lon } = await getCityNameInformations()
@@ -183,8 +157,6 @@ function Home() {
                                     src={`/assets/WeatherIcons/${weatherInformations.today.icon}.svg`}
                                     alt={`${weatherInformations.today.description}`}
                                 ></img>
-
-
                             </IconColumn>
                             <InformationsColumn>
                                 <InformationsColumnItem>
